+++ App.tsx
@@ -26,10 +26,12 @@ import {
   Heading,
   Radio,
   Tiles,
+  PeopleTabs,
   ThemeProvider,
   IconToggle,
   PillToggle,
   TextInput,
+  DataBlock,
 } from './src';
 import Icon from './src/modules/icon/icon';
 
@@ -80,6 +82,8 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
     React.useState<boolean>(false);
   const [selectedHorizontalRadio, setSelectedHorizontalRadio] =
     React.useState<number>(2);
+  const [selectedPeopleTabItem, setSelectedPeopleTabItem] =
+    React.useState<number>(1);
 
   const pickerOptions = [
     {label: 'Option 1', value: 1},
@@ -323,6 +327,12 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
     <View style={{width: 12, height: 12, backgroundColor: 'red'}} />
   );
 
+  const peopleTabsItems = [
+    {label: 'Tab 1', value: 1},
+    {label: 'Tab 2', value: 2},
+    {label: 'Tab 3', value: 3},
+  ];
+
   return (
     <ThemeProvider>
       <SafeAreaView>
@@ -425,6 +435,106 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
                   </Text>
                 </Accordion>
               </Row>
+            </Section.Content>
+          </Section.Wrapper>
+          <Divider />
+
+          {/* Data Block */}
+          <Section.Wrapper>
+            <Section.Title>
+              <Heading variant="h2">Data Block</Heading>
+            </Section.Title>
+            <Section.Description>
+              <Text variant="caption">
+                List like component that renders a label on the right and a
+                value or a React component on the left.
+              </Text>
+            </Section.Description>
+            <Section.Content>
+              <Row>
+                <DataBlock items={pickerOptions} />
+              </Row>
+              <Row>
+                <DataBlock
+                  items={[
+                    {
+                      label: 'Doggo 1',
+                      jsx: (
+                        <View
+                          // eslint-disable-next-line react-native/no-inline-styles
+                          style={{flexDirection: 'row', alignItems: 'center'}}>
+                          <View>
+                            <Avatar uri={avatarUri} radius="small" />
+                          </View>
+                          <View
+                            // eslint-disable-next-line react-native/no-inline-styles
+                            style={{marginLeft: 10}}>
+                            <Text variant="caption">Duffy The Pup</Text>
+                          </View>
+                        </View>
+                      ),
+                    },
+                    {
+                      label: 'Simple Value',
+                      value: 'Lookie Here!',
+                    },
+                    {
+                      label: 'Doggo 2',
+                      jsx: (
+                        <View
+                          // eslint-disable-next-line react-native/no-inline-styles
+                          style={{flexDirection: 'row', alignItems: 'center'}}>
+                          <View>
+                            <Avatar uri={avatarUri} radius="small" />
+                          </View>
+                          <View
+                            // eslint-disable-next-line react-native/no-inline-styles
+                            style={{marginLeft: 10}}>
+                            <Text variant="caption">Buffy The Pup</Text>
+                          </View>
+                        </View>
+                      ),
+                    },
+                    {
+                      label: 'Simple Value',
+                      value: 'Lookie Here Again!',
+                    },
+                  ]}
+                />
+              </Row>
+            </Section.Content>
+          </Section.Wrapper>
+          <Divider />
+
+          {/* People Tabs */}
+          <Section.Wrapper>
+            <Section.Title>
+              <Heading variant="h2">People Tabs</Heading>
+            </Section.Title>
+            <Section.Description>
+              <Text variant="caption">
+                Tabs component used with person record and similar components.
+              </Text>
+            </Section.Description>
+            <Section.Content>
+              <Row>
+                <View
+                  // eslint-disable-next-line react-native/no-inline-styles
+                  style={{
+                    backgroundColor: 'green',
+                    padding: 12,
+                  }}>
+                  <PeopleTabs
+                    items={peopleTabsItems}
+                    onChange={item => setSelectedPeopleTabItem(item.value)}
+                    selected={selectedPeopleTabItem}
+                  />
+                </View>
+              </Row>
+
+              <Text variant="caption" muted>
+                The green background is only used for contrast.
+              </Text>
             </Section.Content>
           </Section.Wrapper>
           <Divider />